Floating point support

+++ include/ast.h
@@ -2,5 +2,5 @@
 #include "stack.h"
 
 Node createTree(Stack *);
-int executeNode(Node *);
+float executeNode(Node *);
 void printTree(Node *);
